[platform] specify URL for custom built-in plugin repository via system property instead of using ApplicationInfo.xml (IJPL-204)
This feature isn't used for public builds, only for nightly builds and for IDEs built from sources, and the actual URL is computed by build scripts so it's better not to have an attribute for that in each *ApplicationInfo.xml.

IntelliJ IDEA has been migrated to use the new property.

GitOrigin-RevId: 751cdab46a9860c018d277a066307835e167d08d