[platform] fix ModuleUtilCore::findModuleForPsiElement for file located under module and library roots (IDEA-330034)
If a file is located under a module source root and under a library classes root, 'isInLibrary' will return 'true' but 'getOrderEntriesForFile' won't return 'LibraryOrderEntry' for it. This should be probably fixed during reimplementation of 'getOrderEntriesForFile' (IDEA-307389), but for now let's implement a simple and safe fix: just don't return 'null' in such cases and let 'getModuleForFile' find the proper module.

GitOrigin-RevId: 77bea4264dcfe03d47b9ace62f4aa8054767460e