KTIJ-26112 [kotlin][api-watcher] Create `KtSourceModuleByModuleInfoForOutsider` for deleted files
When file is deleted from the FS, `createNonLocalPsiFile` is called
with `localFile = null`. The fake file is still marked as outsider,
but `getOriginalFileUrl` for it is `null`.

Because of that, `KtSourceModuleByModuleInfoForOutsider` was not
created for such "removed" files, because `getOutsiderFileOrigin`
was just returning `null` for them.

This, in turn, caused exception during KtSymbol pointers restoration,
because `canBeAnalysed` check in `KtAnalysisScopeProviderImpl` was failing.

With this commit, `KtSourceModuleByModuleInfoForOutsider` is created
whenever the file is marked as outsider, even if the original file
is unrecoverable for some reason.

Also, moved marking file as outsider to `markAsOutsiderFile`
function, so all the "accessors" for that information on
Kotlin side are grouped in a single place.

Also, add explicit nullability annotation to the
`GlobalSearchScope.fileScope` function, so it's clear
that it accepts nulls.

^KTIJ-26112 Fixed

GitOrigin-RevId: 2d2c0f74bded68b9896fcd69009b65787c6266db