IDEA-330663 Replace JBInsets.key with a supplier and rewrite addInsets
Instead of just adding scaled insets, save copies of original insets,
and then use them to calculate the unscaled insets for the result
on every update().

To do so, replace the UI Defaults key with a more generic supplier.

Now the resulting insets will reflect all updates of the insets being added.

GitOrigin-RevId: ced78822f46d379b45e74bc96360fbce68308f22