IDEA-330663 Actually use the value from UI Defaults when creating JBInsets
The recent changes enabled JBInsets to update their values from UI Defaults,
but only when update() is called explicitly. It doesn't always happen,
as this method is only used in a few places for now, so the insets would
often just stick to the hardcoded default.

FIx by rewriting the constructor in such a way that it only performs initialization
now, no scaling or any other logic. This allows us to delegate the actual
calculations to public API methods, such as create(), which now calls update()
after creating the object, thus retrieving the value from UI Defaults or from
another supplier.

GitOrigin-RevId: 66e5fd57614da2b98d6fb5ac0096c48273bc643f