IDEA-319881 Get rid of slow ops in AbstractProjectViewPane.MyDragSource.canStartDragging
To get rid of PSI pointer resolve, we check the same thing in a different way.

Previously, it would return true if there are PSI elements selected (doesn't matter which ones)
or if there are any modules selected. Implement this with instanceof check
instead of pointer resolve. This can potentially break some edge cases when
a node doesn't extend AbstractPsiBasedNode but still contains some PSI elements
or, in a similar fashion, a module node doesn't extend AbstractModuleNode.
If there are any such cases (which is unlikely), they'll need to be handled separately.

GitOrigin-RevId: 0cd1b5fe84adac37e13a8288b9a7b0d7c8719309