[gradle] [IDEA-330949] Attach transformed artifacts components on project import if components exist in Gradle cache.
Right now Gradle import attaches components of a regular java artifact (like sources and javadocs for generic JARs). Components of transformed artifacts, such as sources and javadoc for Android AARs, will not be attached even if the original sources of AAR library exist in Gradle cache. The reason for this is the behavior of GradleProjectResolverUtil, since we search for the sources and javadocs for the artifact in the adjacent path, and since the transformed artifacts are stored in different folders, we cannot find the correct sources and javadoc for the artifact. The solution introduces GradleLocalCacheHelper that can search in the Gradle cache based on the artifact coordinates

GitOrigin-RevId: ebdabae7884f31e32919958275595fb239ed65ff