[gitlab] Add versioning to GQL fragments
The incredibly confusing semantic is now:
A fragment in a specific directory is the one used until that
version. The fragments root directory reads as 'until the
latest version of GitLab'.

Reasoning is that this makes it so that the root directory
reflects the latest versions and thus the gitlab.com API.

fixup! [gitlab] Add versioning to GQL fragments

GitOrigin-RevId: 55f5e5581789a3eb425e121b14f56928417dc050