IDEA-319883 Move package dependencies tree model to background
Create BackgroundTreeModel to use in combination with AsyncTreeModel.
An alternative would be to migrate existing models to support InvokerSupplier,
perhaps by using the ubiquitous StructureTreeModel, but since the model
is created using an extension point, it wouldn't be so easy (though it's
only implemented in Rider, so still possible).

Delegate everything to the original model except listeners, as those
are added early, so use indirect delegation for them: have our own
list and delegate lazily.

The code that used to expand/select tree nodes on init had to be converted
using TreeUtil, as the model isn't initialized yet when the panel is created,
so we need to perform this in background too.

To get rid of some slow ops in the renderer, add PackageDependenciesNode.update
method and call it whenever appropriate. Right now it's called
whenever a node is retrieved from the backing model, but it still needs
some research: perhaps we need to subscribe to some events and update
when those happen.

GitOrigin-RevId: be267409f266da174e847c1009cc012892079d1a