[terminal] IDEA-331027 Allow navigation between blocks when command is running
Send key events to the shell only when no selection or running block is selected.
But anyway, it is impossible to move to the previous block using UP key when a running block is selected, because we need to send it to the shell. So, other blocks can be selected only by mouse.

GitOrigin-RevId: 0e4ba309385ce9656821e286497c30a74f8ce2d4