[workspace model] IDEA-321120 Fix file name calculation for the module-related entities
Before this fix, we were taking into account only FacetEntities as additional entities that need to be serialized to the module-specific file. However, after the introduction of the new API for modules' own settings instead of Facets I forgot to add new custom entities to the calculation of the module-specific files. So, as a result, the manually added facet was not serialized on disk.

GitOrigin-RevId: 94610596a26839ea9b97d6b4b84da2532f974633