IDEA-321006 Guarantee icon path patcher from ExperimentalUIImpl is called last
The order of installing icon path patchers is not guaranteed, because they are called in parallel.
If the patcher from ExperimentalUIImpl is added before an external patcher, it doesn't allow the external patcher even try to patch the path. By adding ExperimentalUIImpl patcher to a separate array the correct order is guaranteed.

GitOrigin-RevId: 97abde92a52b0f5d50cff908eec49b993bbb7cd2