IDEA-319883 TreeExpansionMonitor.restoreAsync
Use TreeUtil async API to restore the tree because with AsyncTreeModel
it's impossible to do otherwise: the model is simply not initialized straight away.

Since we only do it on a full model rebuild, cancelling the previous restore
is not necessary, as it's cancelled when the previous model is disposed.
However, we still check if the model is the same, as otherwise we don't
need to do extra work and also don't need to unfreeze at the end because
there's another restoreAsync running.

GitOrigin-RevId: bad47ce7e4a9a20e0734d9f4d6a7ea39592920a5