[startup] include activities from new modular loader to the startup performance report (IJPL-128)
When the new approach to start the IDE is used, the initial classpath passed to JVM command line includes platform-loader.jar only. The full classpath is loaded in code before com.intellij.idea.Main.main is called, so we need to record timestamps of the initial activities to get correct startup performance report for the new approach.

GitOrigin-RevId: 5b8b2cc138a32ca1a29ff355f4dbbae12e3f0cbc