[rdct] refactoring: move ThinClientPathCustomizer to platform and rename it to PerProcessPathCustomizer
This is needed to allow reusing this functionality in other products (IJPL-223). Also, it makes it simpler to start JetBrains Client using the new modular approach (IJPL-128), because an implementation of PathCustomizer is created during the early phase of startup, so its module must be somehow added to the classpath.

GitOrigin-RevId: 18f45d60f2e511a115392ba77d1b983b3a370c88