[kotlin] k2: don't search enum inheritance in foreign languages (KTIJ-26876)
If enabled, `JavaDirectInheritorsSearcher.calculateDirectSubClasses` finds kotlin enum constants as light classes (and k1 relies on that). Based on those classes, light methods are found. But for k2 these results are redundant, because direct kotlin search is performed as well.

Regular Kotlin inheritors are not identified by the `JavaDirectInheritorsSearcher` due to their absence in the `JavaSuperClassNameOccurenceIndex`.

GitOrigin-RevId: 6ba01326975297d73ed88085154fd9706dec2b50