IDEA-330211 Renamed ThemesListProvider to ThemeListProvider
The class and all references to 'ThemesListProvider' have been renamed to 'ThemeListProvider'. This includes not only the class name itself but also all places within the code where this class was utilized. The reason behind this change was clarity and language consistency. The word 'Themes' was identified as potentially leading to misunderstandings due to its plural form. Renaming it to 'Theme' aligns the language used within the code to common conventions and improves readability.

GitOrigin-RevId: f3fc0b8c6a4ee46dbd51513994aa17fc9a5a105b