FIR/UAST: introduce fake suspend continuation parameter for fake method
Its type is `Continuation<? super suspend-fun-return-type>` that
requires FE-specific ways to resolve Continuation; build a type with the
containing suspend function's return type as a type arguemnt; and
convert that to PsiType. To that end, a new API is added to the base
resolution service.

Handling fake method for suspend function is implemented in general,
including adding that fake suspend continuation parameter w/ a proper
type and nullity and setting @Nullable Object as a return type.

^KT-61544 Fixed
^KTIJ-26853 In progress