[gitlab] Make snippet creation possible with no logged in accounts (IDEA-327321/IDEA-330962)
When no user login information can be found whatsoever,
the user is prompted with a warning that presents an
action to login a new account.

Also, this commit disables the OK button when it is completely
impossible to post a snippet, rather than waiting for
verification, as verification is already done and a warning
is shown.

GitOrigin-RevId: a5e87f3cb4f1cede012190f0e89eeea6d349697d