[gitlab] Snippets: add error line to inform user of missing contents (IDEA-327323)
If no contents can be used, the OK button is disabled and an
error line is shown to let the user know no contents are usable.

If some contents cannot be used, a warning is shown to let the
user know some contents are empty, with the files that are
empty.

GitOrigin-RevId: 9fac329f1dd0df830ba03b99cc6fef019dd3e19f