vcs: use ModalityState.any() for 'setErrorStripeRenderer'
'ModalityState.defaultModalityState()' works, as 'EditorTextField' editor is created on the first 'getPreferredSize' call,
that currently happens only after entering the modality for DialogWrapper. But this is not a reliable constraint.

'ModalityState.stateForComponent' is not a better alternative, as components may not be added into hierarchy either.
'EditorComponentImpl' of 'EditorTextField' is added into the hierarchy only after the 'EditorCustomization' are processed.

GitOrigin-RevId: c3c51fd03a7bfb1219d429ac5085036c1c8d232d