[IDEA] Add point translation method to InputHandler class (fix IDEA-331481)
This commit adds a new method to the InputHandler class named 'translatePoint'. The purpose of this new method is to allow for changes in inlay points in the model if it differs from the real presentation. This has been implemented to handle differences in inlay points in the CodeVision Inlay renderer and editor mouse motion listener class.

GitOrigin-RevId: 9a6a32eac04e6733defd5ec3358a1e7e7d9d9cef