IDEA-326435 Hide Got It tooltips with invokeLater
This is needed because otherwise things get very messy if
the component being hidden also existed in the layered pane.
What happens is that a hierarchy callback is triggered
from JLayeredPane.remove(), while it's still in process of removing
the component, this callback causes the Got It tooltip
to be removed as well, and when control gets back to
removal of the component, then indices are no longer valid,
and everything goes wrong.

Invoking this in an invokeLater turns both removals
(the component and the tooltip) into two separate actions,
so they no longer interfere with each other.

GitOrigin-RevId: 635adf0de840cc39a3a0484679881912dd098165