IDEA-331504 Update UI theme handling to improve performance and clarity
This commit refactors the way UI themes are handled in the IDE, moving away from the LafManager to the UiThemeProviderListManager to get a list of available themes. This aims to improve performance by avoiding unnecessary UI updates and clarifies the code by consolidating theme handling in UiThemeProviderListManager. Also, a function parameter is added to loadFromJson in UITheme in order to decouple it from LafManager, further increasing clarity and testability. The QuickChangeLookAndFeel class is updated to accommodate these changes. Minor formatting and cleanup changes are also included to improve code readability.

GitOrigin-RevId: 6510a7183d18528673fb49942984199fd768aa5a