[build scripts] don't include descriptors for unrelated libraries into runtime module repository (IJPL-109)
A module descriptor for a library which is a part of another library should be included into the module repository only if some include module depends on that library. Otherwise, unrelated libraries may be included into the repository (e.g. 'eddsa' library from 'fleet.plugins.ssh.frontend' which is a part of SSHJ library).

GitOrigin-RevId: 7d89f552638775f3035d468dd67a22c53aa564fa