[startup] allow to always load *ApplicationInfo.xml from resources (RDCT-234)
Information about the main IDE is injected into 'getAppInfoData', so in order to start JetBrains Client from a distribution of another IDE, we need to load data from the corresponding *ApplicationInfo.xml file from the resources. This is needed until we can get rid of ApplicationInfo.xml for JetBrains Client completely (RDCT-460).

GitOrigin-RevId: 654e9fbebe70dbe350c39c9abc605194ed80d91c