[startup] make sure that ApplicationInfo is loaded before ApplicationImpl is instantiated
ApplicationImpl's constructor loads ApplicationInfo in its registerFakeServices method, so if we don't wait until it's loaded explicitly, time to load it will be included in "app instantiation" phase. Also, if the new modular loader is used (IJPL-128), it may fail to load ApplicationInfo if the main classloader isn't fully initialized.

GitOrigin-RevId: 31412370247458707b0fb9f38d6e08dea2a223d7