[java-analysis] ClassDataIndexer: cache bytecode analysis gist data for virtual files
It appears that ourGist.getFileData() reads the whole class data from the disk every time we ask equations for a method/field. If the class is huge, it's deserialized completely, then equations for a wanted method are returned, and the result is discarded. So, if we need to know contracts for N methods of a single class, we will deserialize the class data N times, which is really bad for huge classes.

Tried to read contracts for all methods of ThriftHiveMetastore$Client.class in a loop (from Apache Hive). Before this change it took ~50 seconds, after this change it takes only ~5 seconds.

Possible solution for IDEA-327595 Do not build `BytecodeAnalysisIndex` for huge class files

GitOrigin-RevId: 1639432890699fc67ce30696fcb8a7d3cca3537d