IDEA-326435 Add speed search shortcut description to the menu
Turned out to be trickier than expected. The text is not standard,
so BegMenuItemUI needed to be provided the alternative text.

But the existing implementation (ActionMenuItem.getFirstShortcutText)
simply called KeymapUtil.getFirstKeyboardShortcutText
with no way to override it.

There's no easy way to move that code to AnAction and/or Presentation,
as its module doesn't depend on the module containing KeymapUtil.

So the best I could think of is to implement it with a client property in Presentation,
but leave the default in ActionMenuItem.

GitOrigin-RevId: 309e14aea93f18da4e927f6103498ec7c53d77ad