IDEA-330342 Remove rowHeight-related Linux hacks
With the new DefaultTreeLayoutCache it causes the project view
to jump up and down as node sizes are recomputed, which
happens quite often.

This may break UI in certain Linux environments,
but should only affect Linux. We'll revert this if anything
goes wrong, but since it's been a while since the last attempt
to fix this, the only way to fix this for good is to try again,
as many things may have changed since then: Linux environments
have better scaling support now and the platform evolved a lot too.

GitOrigin-RevId: 0a9f5a700489f302c5d7f08c6a2207e238b2951f