[search] IDEA-320341 Refactor: add an ability to specify layer for highlightings
It is needed in terminal search to use different background for the current occurence and to show highlightings on top of the selection.
Using the markup model directly instead of `HighlightManager` (without its feature to remove the highlighter on escape pressed) should be safe, because highlighters are removed when the search popup is hidden from screen.
But I found that it is not working in one of the tests (`InjectedLanguageEditingTest#testHighlightSelectionInWholeEditor`), so I added the logic with hiding the highlightings to the `EscapeHandler`.

GitOrigin-RevId: 4abc5702c9001eb521c27de1ab86cb5850c965b6