LLM-752 WIP Add semantic search functions to the chat
This commit includes the following changes to the codebase:

- Introduction of the 'SmartChatEndpoint' abstract class in order to define functions within the chat
- Creation of two functional endpoints: 'SemanticSearch' and 'MeaningOfLife'
- Inclusion of their functionalities to 'ChatSessionImpl' through the 'smartChatEndpoints' list
- Update of ActionEmbeddingsStorage and DiskSynchronizedEmbeddingsStorage to extend the abstract class 'AbstractEmbeddingsStorage' for code consistency
- Some unused code were removed from 'SuggestRefactoringChatIntention.kt', 'SmartChatMessageHandler.kt', and 'FunctionCallsMessage.kt'

These updates are to facilitate the development of the functionality of smart chat features, and to keep a structured approach on the implementation of each function while separating their logic from other components. The 'SemanticSearch' function allows users to search items within their projects using natural language queries, whereas, 'MeaningOfLife' is a tribute to popular culture and returns a consistent computational answer of '42'.

GitOrigin-RevId: 041309428e9c4186a172a8ea14631acbb2dbdded