IDEA-330211 Refactor UI theme parsing code into separate file
A new file, `uiThemeParser.kt`, has been created for the function parseUiThemeValue and related private parsing functions. They were previously in `UITheme.kt` but have been moved to simplify and improve the structure of the codebase. This makes the code easier to read, manage, and debug, as functionality related to parsing UI themes is now encapsulated in a single file.

GitOrigin-RevId: 68471659f7251380fd21725f7292bab032d8afa8