IDEA-330211 Refactor color parsing in UI theme parser
Simplified color and integer value parsing in uiThemeParser.kt file using 'let' function for null-checking instead of traditional conditional checking. This enhances readability and code neatness. Additionally, improved the parsing of color to include alpha channel offset and ensured that the color is properly defaulted even if the hex color code starts with '#'.

GitOrigin-RevId: bd1080592ad8da09c864cac2950fd58a7fd3dd2d