IJ-CR-114421 Merge speed search actions into a single context-dependent action
Now that we replaced custom shortcut texts with custom suffixes for them,
we need to have a single action, because it must both have a shortcut
and be present in the tool window menu, which wasn't the case before,
as it were two different actions.

Fix by removing the tool window specific action and instead implement
the same functionality in TreeSpeedSearchAction. Now it works for
the context element if it has speed search, but if it isn't, it looks for
a tool window decorator and works for the top-left-most tree instead.

To deal with the focus issue, just check the action place.

GitOrigin-RevId: 9f80934409c83fe59f664a1dc6ef68906ecbc5c5