IDEA-326435 Fix speed search action picking the wrong component
Some tool windows (e.g. some Profiler tabs) have several speed search
ready components, but choose one of them to be the main context component.
This caused SpeedSearchAction to bypass component lookup
and invoke the action on the context component instead.

Fix by checking if the action is invoked with the keyboard shortcut,
and if not, always perform full search without checking the context
component first.

GitOrigin-RevId: 3666fba061113b279bbca377c79777a14322b449