Allow bundled K2 Compose compiler
In order to prevent binary compatibility issues,
KotlinK2BundledCompilerPlugins ensures that only "bundled" Kotlin
compiler plugins are loaded in the IDE.

The Compose compiler plugin is a corner case. It is not listed in
KotlinK2BundledCompilerPlugins, yet we ship it in Android Studio and
guarantee its compatibility with the bundled Kotlin compiler.

For now we will patch the Kotlin plugin to allow our bundled Compose
compiler to run. The long term plan is to move the Compose compiler
plugin directly into JetBrains/kotlin, or otherwise create an
extension point for bundled plugins to substitute additional compiler
plugin jars.

Note that findBundledComposePlugin() is needed because the jar
location may be different in dev builds vs. release builds.

Bug: 293653361
Test: covered indirectly by K2 Compose tests
Change-Id: I84190a46184995e1a21447da017f38813a9ff3e2