IDEA-331273 Log FrameWrapper bounds change by API
This is needed to trace every attempt to move/resize a windowed tool window
from the user code. Note that it doesn't cover outside changes, from the JVM,
originating from the OS or the window manager. But at least we can say that
if bounds did change, but there's no corresponding trace log message,
then the change must have originated from the environment and not our code.

GitOrigin-RevId: a53a2f96f592c67be35b5f4f960b1eebc000c4f0