GTW-5734 Fix keyboard events don't work in Git Branches popup opened form toolbar
In toolbar there is a custom component created for Git Branches action.
The component is created without any parent - for the purpose of converting component ot BeControls.

However, the component is used as an owner to popup it creates. And popup can't request focus unless it has the visible window as an owner (see `java.awt.Window#isFocusableWindow`)

To fix that, I'm adding a foster parent to custom component - to make sure it have a visible window as a parent

GitOrigin-RevId: 550c8dc9d47315a77b12f0a3e8e27df6ee10bd7d