(IDEA-325291) ML in SE: enable semantic search experiments in Search Everywhere tabs
By default, semantic search is disabled in all tabs. If a user is internal or has an EAP build with a corresponding experiment group, we manually enable the semantic search on startup BUT only if it was not disabled manually. The fact if it was disabled manually is persisted on disk among other settings.

GitOrigin-RevId: 47481679792cdf1df8649d9ecd4b715b6625a62a