CWM-8569 Make ShowPluginManagerAction Frontend only
Plugin manager should be opened locally on a frontend.
It was a frontend action before global actions refactoring, so I'm just reverting it to its old behaviour.

Right now the action is executed on a backend and plugin dialog isn't able to convert to BeControls in CWM (so it either doesn't show, or shows broken layout)
In RD the action shows luxified settings which are differ from the one, opened locally

(cherry picked from commit 465cb17bcdea4dc0721d4866afa1973e6329816f)

IJ-CR-115096

GitOrigin-RevId: b46dc4b4685df45c985a42d069816d9652506491