[rdct] run JetBrains Client from existing distribution (RDCT-163)
If an IDE is asked to run JetBrains Client, the host version is equal to the version of the current IDE and "rdct.use.embedded.client" registry option is enabled, the IDE will start the client process using files from its own distribution instead of downloading it. This works only if runtime module repository descriptors are bundled with the IDE (it isn't enabled by default yet, see RDCT-517).

GitOrigin-RevId: 891d44d1a88d9a7e7cec14614ca575fed0f0c618