[vfs] IJPL-173: store allocatedRecordsCount directly in a header field
We cached .allocationRecordsCount in a class field before. Directly reading/updating an appropriate header field is more resilient: less chance to lose an update (allocated record) if app is terminated -- fewer 'fileId ... is out of allocated id range ...' errors.

The downside is: it is probably marginally slower. But I doubt it is noticeable -- btw, VFS is reads-dominated.

GitOrigin-RevId: 5d9004015e199960b840d2336c5a302bd8d0392e