Refactor tree structure and selection logic in jewel library
Code has been implemented that simplifies tree structure handling and improves logic related to the selection of tree elements.

`idPath` methods, which used to get path IDs, have been replaced by simpler and more efficient `id` calls. This reduces unnecessary iterations and simplifies the code.

The selection functionality in KeyBindingAction and BasicLazyTree has also been streamlined. Checks for element clicked now rely on the ID rather than the ID path. This again reduces non-essential operations.

In TreeState, the `openNodes` collection is now a mutable state of Sets. This allows better performance when querying if a node is open, which was previously an 'O(n)' operation.

Code related to the user interface in BasicLazyTree has also been refactored for simplicity.

For instance, the `itemsIndexed` function is now used instead of `items`, which allows the use of index directly, improving readability and maintainability.

This commit aims to simplify, improve the performance of the code, and enhance the user experience interacting with tree structures.

GitOrigin-RevId: 9d14b4de3e1f228fca381712b569365e7751ef0b