[kotlin] Consistently filter out elements from parent context.
In incremental analysis, elements contained within the reanalyzed
element should not be looked up in the parent context. PsiElements
are destructively updated when the code changes, and therefore
old and incorrect information could be fetched from the parent
if allowed.

This change extends the filtering from only happening in
getSliceContents to also being in place in getKeys and get.
This way, the stacked binding context consistently disregards
elements that are contained within the reanalysis scope
from the parent context.

^KTIJ-26856 Fixed.

GitOrigin-RevId: 4a9ab289c68ada776823d2aabaf1731c15a0cbca
(cherry picked from commit 25121ff89d69f8a52657613ffd92c44675f70363)

Bug: b/295257198
Test: ComposableCompileTest
Change-Id: I505ea0ff0064bc5bb7f5dc77c7f4070ddfb2edce