[code style][CLI formatter] do not rely on `Application.isHeadlessEnvironment` to run computations synchronously
For CLI formatting, we want to compute code style settings synchronously, i.e. block the EDT. Previously, we used `Application.isHeadlessEnvironment` to detect such case, now there is a special flag for it.

The problem is `isHeadlessEnvironment` will also include many other scenarios, for example build tests. Additionally, `EditorCodeStyleSettingsModifier.modifySettings` uses `runBlockingMaybeCancellable` which logs an error if called on EDT. In CLI formatting, this error is logged, but apart from that, everything works. In build tests however, it will result in a failure of the test.

GitOrigin-RevId: 3de526ecccc24e3078b9b28d679d6bc6f7ea9d78