[K2] Port NamingConventionInspections to K2
Port the inspections in NamingConventionInspections to K2. The bulk of
these inspections are PSI-only, but FunctionNameInspection requires
resolve, so it has been split into a base class to handle the PSI-only
portions of the inspection, and two concrete implementations (one each
for K1 and K2), to handle the analysis.

To more cleanly distinguish between ported and unported inspections, add
an inspections-k1 module to accompany inspections-shared and
inspections-k2.

^KTIJ-22913 fixed