[rdct] build scripts: update reference to license classes while scrambling intellij.cwm.guest module (RDCT-535)
In order to allow running JetBrains Client from an IDE's distribution, we need to include `intellij.cwm.guest` module to CWM plugin distribution. But this module contains classes (RegisterAction and ThinClientLicenseHelper) which refer to internal license-related classes from the platform. These classes are renamed during scrambling, so in order to avoid breaking references to them, this change introduces PluginLayout.allowUsingScrambledPlatformClasses option, which allows passing paths to the platform JARs to the scrambling tool along with the change log from platform's scrambling process to ensure that the platform classes will be renamed to the same names during plugin's scrambling process.

GitOrigin-RevId: 2c2edb5251331704448f798382738c072a9545c3