[util] fix race condition in ClasspathCache (RDCT-538)
Since ClassPath.findResource may be called concurrently with changes in ClasspathCache, we need to ensure that it stores values of allUrlsWereProcessed and lastLoaderProcessed before accessing cache, otherwise it may happen that these values are changed while findResource is iterating over (old value of) loaders, and this will result in "Resource not found" exception. Similar problem may happen in findClass.

Also, resourcePackageCacheGetter must be updated only when newResourceMap is fully initialized to ensure that other threads won't see it in partly initialized state.

GitOrigin-RevId: 9da658d831464d19eeedc270963ba8276695ac97