KTIJ-27174 [K2][scripts]: resolution candidates are ignored
There might be several candidates during the resolution process. To pick
only one of them, [1] is used. This function is based on a mapping between
KtModule (corresponding to the candidate's PSI) and its precedence rank.

The problem is that the mapping keys are calculated from one context while
searching keys from another context. Thus, we deal with instances of different
classes ([2] and [3]). Given the keys are not comparable, no candidate is picked,
resolution fails.

As a fix, the context [4], where representatives of [3] were instantiated, was
broadened. From now on, the analyzed script file is taken into consideration and
instances of [2] are created instead.

-----------------------------------------------------------------------------
[1]: LLFirSelectingCombinedSymbolProvider#selectFirstElementInClasspathOrder
[2]: ScriptDependenciesInfo.ForFile
[3]: ScriptDependenciesInfo.ForProject
[4]: ProjectStructureProviderIdeImplKt.calculateKtModule

^KTIJ-27174 fixed

GitOrigin-RevId: 16a2b33c4fdf78b7deaea851919a0bc744940f52