IDEA-320078 refactor code in FormsParsing.java for simplicity
This commit removes some unnecessary code from the FormsParsing.java file. It removes unused import and simplifies the usage of the InputStream object with try-with-resource. The removed code mostly had no effect, hence the changes are made to improve readability and simplicity of the code. Furthermore, some functions that were not producing any effect were also removed, which contributes overall towards cleaner code.

GitOrigin-RevId: e450dd126a8e31888846bbe25fd54ad4fd297b7b