IDEA-320078 Refactor XMLEntityResolver for better approach and consistency
Changed Hashtable to HashMap for storing entities in XMLEntityResolver class. Hashtable was replaced with HashMap because it's more modern and provides better performance. Also, made the class and its default constructor package-private to adheres to the principle of least privilege. Moreover, the method openExternalEntity has been made static as it does not depend on any instance variables.

GitOrigin-RevId: b1ff545afaf4f08bde5dcd92d5ef07c1db3df899