IDEA-320078 Refactor StdXMLBuilder class to improve code quality
Removed unused imports and replaced Stack with Deque for better performance and readability in StdXMLBuilder class. The Stack class in Java is considered as a legacy class and has been replaced by modern equivalents like the Deque interface which offers better performance. Furthermore, replaced StringBuffer with StringBuilder for better performance when dealing with string concatenations in a single-threaded environment.

GitOrigin-RevId: 4afe6c1de67667ea47146b2b8782410b8895dcc4