[platform] don't add listener to IdeEventQueue during initialization of RunAnythingAction class (IDEA-331510, GTW-6247)
It may cause performance issues and lead to exceptions if the class is firstly accessed from a background thread. Values of SHIFT_IS_PRESSED and ALT_IS_PRESSED static fields are used inside RunAnythingPopupUI, so they are replaced by instance fields in that class. The old static fields and RunAnythingPopupUI.getExecutor static methods are left as deprecated for compatibility with external plugins.

GitOrigin-RevId: 92925ef6001bc0ac644377e81bcd5fa2a372b8bb