[rider] Some modifications to enable the render doc functionality.
Includes a new application service and a new project service for overrides in Rider, also some class visibility/virtuality tweaks for the same reason.

The change to the DocRenderModel.kt is due to RdGen parsing the `sink` declarations to be sinks relative to the backend for Rider.

The changes to `DocRenderModelSupport`s are required because there is no way to remove those supports from the supports extension point and Rider needs to use an implementation of a `DocRenderItemManager` that is not `FrontendDocRenderItemManager`.

(cherry picked from commit 608dd9337ca37deb2880fa12167f089744de2fc3)

IJ-CR-115346

(cherry picked from commit 21954d9de08729b88514b085ef8c9a29b6b205ce)

IJ-CR-115346

GitOrigin-RevId: 2cfbb8cbbdc02d34a790fb14c55c4519b1e0d354