IDEA-332724: expected actions are missing from search when semantic search is enabled
The problem appear because "Fetch" action is found by both regular and semantic search. When we merge information about those elements, we remove already found one and replace it with a new item with both regular and semantic score. On the UI buffer side, items are considered identical if they are returned by the same contributor and point to the same MatchedValue (`SearchEventsBuffer#areTheSame`). Since we reuse original MatchedValue, this condition is true for original item and merged one and we end up removing both original and merged item from the list.

GitOrigin-RevId: 4711afbde9e5e2527b79840f4c679810a00b0c3d