IJPL-166 Refactor and streamline startup process
Renamed `StartupUtil.kt` to `main.kt` as part of the restructuring of the project bootstrap process. Several methods and routines were moved around, notably `startApp` and `scheduleEnableCoroutineDumpAndJstack`. The idea is to create a more readable and intuitive startup process. Also updated the Logger class references to match the new structure.

GitOrigin-RevId: 6f1c0d2505f4df8ebc74d7f646de72d95c2b5cf7