IDEA-331947 Revert "Revert "[fus, code vision] disable logging of code vision events for now""
This reverts commit e75121b97e22a6d31a71a94f679ce4aa85fff818.

Group "daemon.code.vision" reports data too often - several events per typing.
It should not be enabled before it is properly fixed.
As it is needed to detect problems with performance of providers, it can be reported on file close as duration buckets per provider. It also can be only one event with ObjectEventList and only for applicable providers.

GitOrigin-RevId: dff2bcebb1365d86178ff3199049f64d667ac211