IDEA-333226: simplify the code: there is no need to capture indexing dependencies fingerprint when task is queued
We only need the fingerprint when the task is started, and we don't care that it might have changed while the task was in queue.

Real problem is that obtaining a fingerprint might be a long operation (e.g., when running from sources, we analyze classpath - hundreds of thousands of files). We want to make sure that we know the thread which is used to perform calculations - it should be indexing/scanning thread.

GitOrigin-RevId: 91625bd43e1e5d4be23baf5034505c8770c810ef