[kotlin][debugger] IDEA-318883 ignore module dependencies-related errors coming from diagnostics
These errors come from `IdeJavaModuleResolver` when an evaluation happens in sources jar, and this jar is represented a named module (i.e. it has `module-info.java`). Such modules are not part of a module dependency graph -- we can't rely on the information in `module-info.java` to check dependencies between *classes*. So we can't establish relations between them and other modules, and evaluation within these modules will fail.

One of the solutions if just ignore module-related errors in evaluation. They're just diagnostic reports, not compilation errors, so there are chances the expression will be successfully evaluated. Then again, if we try to evaluate something that's completely unrelated to current context (i.e. call some functions from our homemade module no one depends on), the evaluation fails with an "Unresolved reference" error. (It's desired to be fixed -- it'd be nice if we could evaluate anything we want, but that's another story.)

Code that causes the aforementioned diagnostic issues is available in completion, but will be highlighted as an error.


Merge-request: IJ-MR-115327
Merged-by: Alexander Kuznetsov <Aleksander.Kuznetsov@jetbrains.com>

GitOrigin-RevId: a39534304d2d0d6c37ef1394308711751fee8f7f