[file-history] follow multiple rename points in git-based file history
Previously, when computing file history, only the first encountered commit which added the file was taken into account and further commits were skipped. In this commit, GitFileHistory was modified to collect all commits where the file was added and try to find a rename and follow history for all of them. This allows to find more commits when a file was renamed in two separate branches simultaneously, or when an alternative history was merged.

This implementation does not sort commits topologically, and following different branches can lead to commits being out of order. Since some places in the IDE rely on the commits order, a parameter "fullHistory" is introduced to control this behavior.

IDEA-333216

GitOrigin-RevId: fe6904b9c3c64bad2fe499a77fb62ddb35a018f8