IDEA-332527 Work around macOS race condition with dialog sizes
This is yet another ugly hack to deal with macOS asynchronously
changing Swing component properties. In this case,
when a dialog is pack()ed, it's made displayable, which causes macOS
to start doing its things, one of which directly changes the size
of the dialog from the AppKit thread. This can happen at any moment,
and if it happens after pack() sets the right size, we're screwed.

It's impossible to fix on the client side because the entire call stack
from the AppKit thread is inside Swing internals, no way to override
anything. So instead we keep track of the size set by the API only,
via Component.reshape (which all other methods like setSize, setLocaiton
and setBounds delegate to). While inside show(), we pretend that
no outside size changes can happen. And when the dialog is finally shown,
we compare our tracked size to the actual size, and if it's different,
we resize the dialog.

GitOrigin-RevId: ba4e9bd22e96a3cd619a6f5dbe542aa1737b48c7