[startup] postpone loading lists of disabled, expired and broken plugins
If we load these lists when DescriptorListLoadingContext instance is created, it may lead to exceptions if the main classloader isn't fully configured at this moment (when a new modular loading strategy is used to start JetBrains Client, RDCT-234), because DisabledPluginsState.getDisabledIds() and getBrokenPluginVersions() use ApplicationInfo. Also, postponing loading may improve performance a bit.

GitOrigin-RevId: 0150ffc2fb22c1df82e1eb89889b9ac314e9f5dc