IDEA-332527 Register custom shortcut for components supporting speed search
This takes care of shortcut conflicts between Speed Search and some other action
registered in a broader context, for example, in Settings | Editor | File and Code Templates,
where the same shortcut is used for searching all settings on the root pane level.

This might introduce more regressions when Speed Search takes over when
it shouldn't.

GitOrigin-RevId: 6321073de4f0ec62c9e79cb1283b9f14d46b48f9