IDEA-333521 Use the result from the previous ActionPromoter for the next one
ActionPromoter.promote is designed to return just the actions that need
to be promoted, but some promoters, like, say, EditorTextFieldActionPromoter
are implemented by sorting the entire list of actions. This is needed,
for example, to demote an action instead of promoting it, or to promote
some actions relative to some other actions and not to the entire list.

The problem with this approach is that it only worked for one promoter
implemented that way. This is because the action list passed to every
promoter was not the action list modified by the previous one, but
instead the original list of actions. This effectively led to a "last promoter wins"
scenario. In our case, SpeedSearchActionPromoter would be overridden
by EditorTextFieldActionPromoter.

Fix by passing the modified list of actions instead, wrapping it into
a non-modifiable list for additional safety (in case of a broken promoter
implementation). This is a safe fix, because for the "conventional"
promoters implemented by filtering the list, the order doesn't really matter,
and for "unorthodox" promoters that sorted the entire list instead,
it was already broken. The only thing that may break now is some
actions that relied on this incorrect behavior, but those should be fixed
separately.

GitOrigin-RevId: 8488d83977866210d9e3751a41217029f4a51557