IDEA-331827 Fall back to default layout if no window info when restoring layout
When the layout being restored contains no information for a certain tool window,
it used to keep the old information. Then some fixes were added that at least
made such tool windows invisible, otherwise it was very confusing (IDEA-321129).

However, when it comes to removing stripe buttons, then there's some rem-dev
code that uses a completely empty layout, which would hide all tool windows
and remove all stripe buttons, which broke the test completely.

While it's certainly possible to just fix the test and relevant rem-dev code,
what makes most sense is not to keep the old info, modifying it in
some ways to fix various UX issues, but instead just revert it to the info
from the default layout. This is the least surprising behavior for the user
when the layout is restored, and it covers all these issues as well.

In case even the default layout contains no information about the tool window,
fall back to old logic: use the current window info, but hide both the window
and its stripe button.

GitOrigin-RevId: 0a4e489534bfd3d2aeabe8480a8d12de1255a74d